+++ tools/src/tester/TestRunner.ts
@@ -11,7 +11,7 @@ import type StoryEvaluator from './StoryEvaluator'
 import { type StoryFile } from './StoryEvaluator'
 import fs from 'fs'
 import { type Story } from './types/story.types'
-import { read_yaml } from '../../helpers'
+import { read_yaml } from '../helpers'
 import { Result, type StoryEvaluation } from './types/eval.types'
 import { type ResultLogger } from './ResultLogger'
 import { basename, resolve } from 'path'
